[PHP 8.0] Move more logic to MatchFactory
