+++ rules/Php80/Rector/Switch_/ChangeSwitchToMatchRector.php
@@ -6,7 +6,6 @@ namespace Rector\Php80\Rector\Switch_;
 
 use PhpParser\Node;
 use PhpParser\Node\Expr;
-use PhpParser\Node\Expr\ArrayDimFetch;
 use PhpParser\Node\Expr\Assign;
 use PhpParser\Node\Expr\Match_;
 use PhpParser\Node\Expr\Throw_;
@@ -15,7 +14,6 @@ use PhpParser\Node\Stmt;
 use PhpParser\Node\Stmt\Expression;
 use PhpParser\Node\Stmt\Return_;
 use PhpParser\Node\Stmt\Switch_;
-use PhpParser\Node\Stmt\Throw_ as ThrowsStmt;
 use Rector\Core\Contract\PhpParser\Node\StmtsAwareInterface;
 use Rector\Core\Rector\AbstractRector;
 use Rector\Core\ValueObject\PhpVersionFeature;
@@ -125,15 +123,10 @@ CODE_SAMPLE
                 }
             }
 
-            $match = $this->matchFactory->createFromCondAndExprs($stmt->cond, $condAndExprs);
-
-            // implicit return default after switch
-            $match = $this->processImplicitReturnAfterSwitch($match, $condAndExprs, $nextStmt);
+            $match = $this->matchFactory->createFromCondAndExprs($stmt->cond, $condAndExprs, $nextStmt);
             if (! $match instanceof Match_) {
                 continue;
             }
-
-            $match = $this->processImplicitThrowsAfterSwitch($stmt, $match, $condAndExprs, $nextStmt);
 
             $assignVar = $this->resolveAssignVar($condAndExprs);
             $hasDefaultValue = $this->matchSwitchAnalyzer->hasDefaultValue($match);
@@ -236,65 +229,6 @@ CODE_SAMPLE
         }
 
         return null;
-    }
-
-    /**
-     * @param CondAndExpr[] $condAndExprs
-     */
-    private function processImplicitReturnAfterSwitch(
-        Match_ $match,
-        array $condAndExprs,
-        ?Stmt $nextStmt
-    ): ?Match_ {
-        if (! $nextStmt instanceof Return_) {
-            return $match;
-        }
-
-        $returnedExpr = $nextStmt->expr;
-        if (! $returnedExpr instanceof Expr) {
-            return $match;
-        }
-
-        if ($this->matchSwitchAnalyzer->hasDefaultValue($match)) {
-            return $match;
-        }
-
-        $assignVar = $this->resolveAssignVar($condAndExprs);
-        if ($assignVar instanceof ArrayDimFetch) {
-            return null;
-        }
-
-        if (! $assignVar instanceof Expr) {
-            $this->removeNode($nextStmt);
-        }
-
-        $condAndExprs[] = new CondAndExpr([], $returnedExpr, MatchKind::RETURN);
-        return $this->matchFactory->createFromCondAndExprs($match->cond, $condAndExprs);
-    }
-
-    /**
-     * @param CondAndExpr[] $condAndExprs
-     */
-    private function processImplicitThrowsAfterSwitch(
-        Switch_ $switch,
-        Match_ $match,
-        array $condAndExprs,
-        ?Stmt $nextStmt
-    ): Match_ {
-        if (! $nextStmt instanceof ThrowsStmt) {
-            return $match;
-        }
-
-        if ($this->matchSwitchAnalyzer->hasDefaultValue($match)) {
-            return $match;
-        }
-
-        $this->removeNode($nextStmt);
-
-        $throw = new Throw_($nextStmt->expr);
-
-        $condAndExprs[] = new CondAndExpr([], $throw, MatchKind::RETURN);
-        return $this->matchFactory->createFromCondAndExprs($switch->cond, $condAndExprs);
     }
 
     private function isFollowedByReturnWithExprUsage(Stmt|null $nextStmt, Expr $expr): bool
